+++ app/src/main/java/org/hyperledger/besu/cli/BesuCommand.java
@@ -2736,6 +2736,9 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
         && !dataStorageConfiguration.getHistoryExpiryPruneEnabled()) {
       rocksDBPlugin.addIgnorableSegmentIdentifier(KeyValueSegmentIdentifier.CHAIN_PRUNER_STATE);
     }
+    if (!dataStorageOptions.toDomainObject().getHashPreImageStorageEnabled()) {
+      rocksDBPlugin.addIgnorableSegmentIdentifier(KeyValueSegmentIdentifier.HASH_PREIMAGE_STORE);
+    }
   }
 
   private void validatePostMergeCheckpointBlockRequirements() {