+++ evm/src/main/java/org/hyperledger/besu/evm/worldstate/AbstractWorldUpdater.java
@@ -40,7 +40,9 @@ import java.util.concurrent.ConcurrentHashMap;
 public abstract class AbstractWorldUpdater<W extends WorldView, A extends Account>
     implements WorldUpdater {
 
-  private final W world;
+  /** The worldview associated with this updater. */
+  protected final W world;
+
   private final EvmConfiguration evmConfiguration;
 
   /** The Updated accounts. */