+++ plugins/rocksdb/src/test/java/org/hyperledger/besu/plugin/services/storage/rocksdb/RocksDBKeyValueStorageFactoryTest.java
@@ -311,7 +311,7 @@ public class RocksDBKeyValueStorageFactoryTest {
       final Path tempDataDir, final Path tempDatabaseDir, final DataStorageFormat format) {
     when(commonConfiguration.getStoragePath()).thenReturn(tempDatabaseDir);
     when(commonConfiguration.getDataPath()).thenReturn(tempDataDir);
-    lenient().when(dataStorageConfiguration.getDatabaseFormat()).thenReturn(format);
+    lenient().when(dataStorageConfiguration.getDataStorageFormat()).thenReturn(format);
     lenient()
         .when(commonConfiguration.getDataStorageConfiguration())
         .thenReturn(dataStorageConfiguration);